fix (int)

+++ application/modules/sgr/models/model_121.php
@@ -11,7 +11,7 @@ class Model_121 extends CI_Model {
         $this->load->helper('sgr/tools');
 
         $this->anexo = '121';
-        $this->idu = (int) $this->session->userdata('iduser');
+        $this->idu = (float) $this->session->userdata('iduser');
         /* SWITCH TO SGR DB */
         $this->load->library('cimongo/cimongo', '', 'sgr_db');
         $this->sgr_db->switch_db('sgr');
