+++ application/modules/sgr/models/model_16.php
@@ -12,7 +12,7 @@ class Model_16 extends CI_Model {
         $this->load->Model('sgr/model_12');
 
         $this->anexo = '16';
-        $this->idu = (int) $this->session->userdata('iduser');
+        $this->idu = (float) $this->session->userdata('iduser');
         /* SWITCH TO SGR DB */
         $this->load->library('cimongo/cimongo', '', 'sgr_db');
         $this->sgr_db->switch_db('sgr');
